Add advanced search page styling
Opted to hide the hints on a small device, because it felt a little
cluttered. Wasn't sure how best to phrase the examples without being
too long-winded, so I just went with a literal example for each search
type of how the same query would look in a normal search.